Update of clustering function
Clustering function based on directed graphs. This is a more recent implementation of the directed graph clustering algorithm. It is coded as a straightforward C++ function with four arguments: two for input (data and distance threshold) and two for output (cluster IDs and parent nodes). It returns the number of clusters found.